+++ packages/provider/src/tests/unit/tasks/imgCaptcha/imgCaptchaTasks.unit.test.ts
@@ -174,6 +174,7 @@ describe("ImgCaptchaManager", () => {
 		logger = {
 			info: vi.fn(),
 			error: vi.fn(),
+			debug: vi.fn(),
 		} as unknown as Logger;
 
 		captchaConfig = {
@@ -515,5 +516,42 @@ describe("ImgCaptchaManager", () => {
 		);
 
 		expect(result).toBeUndefined();
+	});
+
+	it("should fail the user if an ip is passed up and it does not match the ip address stored in the commitment", async () => {
+		const userAccount = "userAccount";
+		const dappAccount = "dappAccount";
+		const ipAddress = "1.1.1.1";
+		const dappUserCommitment: UserCommitment = {
+			id: "commitmentId",
+			userAccount,
+			dappAccount,
+			providerAccount: "providerAccount",
+			datasetId: "datasetId",
+			result: { status: CaptchaStatus.approved },
+			userSignature: "",
+			userSubmitted: true,
+			serverChecked: false,
+			requestedAtTimestamp: 0,
+			ipAddress: getIPAddress("8.8.8.8").bigInt(),
+			headers: { a: "1", b: "2", c: "3" },
+			ja4: "ja4",
+		};
+
+		// biome-ignore lint/suspicious/noExplicitAny: TODO fix
+		(db.getDappUserCommitmentById as any).mockResolvedValue(dappUserCommitment);
+
+		const verifyResult = await imgCaptchaManager.verifyImageCaptchaSolution(
+			userAccount,
+			dappAccount,
+			"commitmentId",
+			undefined,
+			ipAddress,
+		);
+		expect(verifyResult.verified).toBe(false);
+
+		expect(logger.debug).toHaveBeenCalledWith(
+			`IP address mismatch: ${getIPAddressFromBigInt(dappUserCommitment.ipAddress).address} !== ${ipAddress}`,
+		);
 	});
 });