+++ packages/provider/src/util.ts
@@ -15,10 +15,15 @@
 import { decodeAddress, encodeAddress } from "@polkadot/util-crypto/address";
 import { hexToU8a } from "@polkadot/util/hex";
 import { isHex } from "@polkadot/util/is";
-import { ProsopoContractError } from "@prosopo/common";
-import { type ScheduledTaskNames, ScheduledTaskStatus } from "@prosopo/types";
-import type { IProviderDatabase } from "@prosopo/types-database";
+import { ProsopoContractError, ProsopoEnvError } from "@prosopo/common";
+import {
+	type IPAddress,
+	type ScheduledTaskNames,
+	ScheduledTaskStatus,
+} from "@prosopo/types";
+import type { IDatabase, IProviderDatabase } from "@prosopo/types-database";
 import { at } from "@prosopo/util";
+import { Address4, Address6 } from "ip-address";
 import type { ObjectId } from "mongoose";
 
 export function encodeStringAddress(address: string) {
@@ -69,3 +74,18 @@ export async function checkIfTaskIsRunning(
 	}
 	return false;
 }
+
+export const getIPAddress = (ipAddressString: string): IPAddress => {
+	try {
+		if (ipAddressString.match(/^((25[0-5]|(2[0-4]|1\d|[1-9]|)\d)\.?\b){4}$/)) {
+			return new Address4(ipAddressString);
+		}
+		return new Address6(ipAddressString);
+	} catch (e) {
+		throw new ProsopoEnvError("API.INVALID_IP");
+	}
+};
+
+export const getIPAddressFromBigInt = (ipAddressBigInt: bigint): IPAddress => {
+	return Address4.fromBigInt(ipAddressBigInt);
+};