+++ packages/procaptcha-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prosopo/procaptcha-react
 
+## 2.6.3
+
+### Patch Changes
+
+- @prosopo/procaptcha@2.6.3
+
 ## 2.6.2
 
 ### Patch Changes