+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prosopo/provider
 
+## 2.8.0
+
+### Minor Changes
+
+- cf59998: Update DB schema
+
+### Patch Changes
+
+- Updated dependencies [cf59998]
+  - @prosopo/types-database@2.7.0
+  - @prosopo/database@2.6.3
+  - @prosopo/datasets@2.6.3
+  - @prosopo/env@2.6.3
+  - @prosopo/types-env@2.6.3
+
 ## 2.7.1
 
 ### Patch Changes