+++ packages/types-env/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/types-env
 
+## 2.6.3
+
+### Patch Changes
+
+- Updated dependencies [cf59998]
+  - @prosopo/types-database@2.7.0
+
 ## 2.6.2
 
 ### Patch Changes